+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/compile/nodecompilation/FragmentParameterValidator.scala
@@ -6,12 +6,15 @@ import cats.implicits.toTraverseOps
 import pl.touk.nussknacker.engine.api.NodeId
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError._
 import pl.touk.nussknacker.engine.api.context.{ProcessCompilationError, ValidationContext}
+import pl.touk.nussknacker.engine.api.definition.ValidationExpressionParameterValidator
+import pl.touk.nussknacker.engine.api.typed.typing.Typed
 import pl.touk.nussknacker.engine.compile.ExpressionCompiler
 import pl.touk.nussknacker.engine.graph.expression.Expression
 import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.{
   FixedExpressionValue,
   FragmentClazzRef,
-  FragmentParameter
+  FragmentParameter,
+  ValidationExpression
 }
 import pl.touk.nussknacker.engine.graph.node.{FixedValuesListFieldName, InitialValueFieldName}
 
@@ -21,12 +24,11 @@ class FragmentParameterValidator(
 
   def validate(
       fragmentParameter: FragmentParameter,
-      fragmentInputId: String,
       validationContext: ValidationContext // localVariables must include this and other FragmentParameters
   )(implicit nodeId: NodeId): ValidatedNel[ProcessCompilationError, Unit] = {
-    val unsupportedFixedValuesValidationResult = validateFixedValuesSupportedType(fragmentParameter, fragmentInputId)
+    val unsupportedFixedValuesValidationResult = validateFixedValuesSupportedType(fragmentParameter)
 
-    val fixedValuesListValidationResult = validateFixedValuesList(fragmentParameter, fragmentInputId)
+    val fixedValuesListValidationResult = validateFixedValuesList(fragmentParameter)
 
     val fixedExpressionsValidationResult = validateFixedExpressionValues(
       fragmentParameter.initialValue,
@@ -35,14 +37,48 @@ class FragmentParameterValidator(
       fragmentParameter.name
     )
 
+    val validationExpressionValidationResult =
+      validateValidationExpression(fragmentParameter, validationContext)
+
     List(
       unsupportedFixedValuesValidationResult,
       fixedValuesListValidationResult,
-      fixedExpressionsValidationResult
+      fixedExpressionsValidationResult,
+      validationExpressionValidationResult
     ).sequence.map(_ => ())
   }
 
-  private def validateFixedValuesSupportedType(fragmentParameter: FragmentParameter, fragmentInputId: String) =
+  private def validateValidationExpression(
+      fragmentParameter: FragmentParameter,
+      validationContext: ValidationContext
+  )(implicit nodeId: NodeId) = fragmentParameter.validationExpression match {
+    case Some(ValidationExpression(expression, _)) =>
+      val ctx = ValidationContext(
+        Map(ValidationExpressionParameterValidator.variableName -> validationContext(fragmentParameter.name))
+      ) // TODO in the future, we'd like to support more references than just "#value", see ValidationExpressionParameterValidator
+
+      expressionCompiler
+        .compile(
+          Expression.spel(expression.expression),
+          fieldName = Some(fragmentParameter.name),
+          validationCtx = ctx,
+          expectedType = Typed[Boolean],
+        )
+        .map(_ => ())
+        .leftMap(_.map {
+          case e: ExpressionParserCompilationError =>
+            InvalidValidationExpression(
+              e.message,
+              nodeId.id,
+              fragmentParameter.name,
+              e.originalExpr
+            )
+          case e => e
+        })
+    case None => Valid(())
+  }
+
+  private def validateFixedValuesSupportedType(fragmentParameter: FragmentParameter)(implicit nodeId: NodeId) =
     fragmentParameter.valueEditor match {
       case Some(_) =>
         if (List(FragmentClazzRef[java.lang.Boolean], FragmentClazzRef[String]).contains(fragmentParameter.typ)) {
@@ -52,7 +88,7 @@ class FragmentParameterValidator(
             UnsupportedFixedValuesType(
               fragmentParameter.name,
               fragmentParameter.typ.refClazzName,
-              Set(fragmentInputId)
+              nodeId.id
             )
           )
       case None => Valid(())
@@ -106,16 +142,15 @@ class FragmentParameterValidator(
 
   private def validateFixedValuesList(
       fragmentParameter: FragmentParameter,
-      fragmentInputId: String
-  ): ValidatedNel[ProcessCompilationError, Unit] =
+  )(implicit nodeId: NodeId) =
     fragmentParameter.valueEditor match {
       case Some(valueEditor) if !valueEditor.allowOtherValue =>
         List(
           if (valueEditor.fixedValuesList.isEmpty)
-            invalidNel(RequireValueFromEmptyFixedList(fragmentParameter.name, Set(fragmentInputId)))
+            invalidNel(RequireValueFromEmptyFixedList(fragmentParameter.name, nodeId.id))
           else Valid(()),
           if (initialValueNotPresentInPossibleValues(fragmentParameter))
-            invalidNel(InitialValueNotPresentInPossibleValues(fragmentParameter.name, Set(fragmentInputId)))
+            invalidNel(InitialValueNotPresentInPossibleValues(fragmentParameter.name, nodeId.id))
           else Valid(())
         ).sequence.map(_ => ())
       case _ => Valid(())